+++ CHANGELOG.md
@@ -41,6 +41,7 @@
  - Dark Mode: MediaElch has gained an experimental dark mode for its _main window_ (#761).
    It can only be set through `advancedsettings.xml` and will be made into a proper
    settings option in future releases.
+ - The TV show search dialog has gained a preview
 
 ### Removed